refactor artifact loading functions for better code organization and readability
