+++ pkg/sbom_test.go
@@ -28,7 +28,7 @@ func TestLoadSBOMFromReader(t *testing.T) {
 	reader := io.NopCloser(strings.NewReader(spdx))
 
 	// Call the function with the test reader
-	doc, desc, sbomBytes, err := LoadSBOMFromReader(reader, size)
+	doc, desc, sbomBytes, err := LoadSBOMFromReader(reader)
 
 	// Check that there was no error
 	if err != nil {
